add compiling ts to js with webpack babel to minimize amount of injected files
