+++ vite.config.ts
@@ -1,7 +1,12 @@
 import { defineConfig, loadEnv } from 'vite';
 import react from '@vitejs/plugin-react';
 import { viteStaticCopy } from 'vite-plugin-static-copy';
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-expect-error
 import tailwindcss from "@tailwindcss/vite";
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-expect-error
+import path from 'path';
 
 export default defineConfig(({ mode }) => {
     const env = loadEnv(mode, process.cwd(), '');
@@ -26,11 +31,15 @@ export default defineConfig(({ mode }) => {
             rollupOptions: {
                 input: {
                     main: htmlFile,
+                    background: path.resolve(__dirname, 'public/scripts/backgroundWorker.js')
+                },
+                output: {
+                    entryFileNames: '[name].js',
                 },
             },
         },
         server: {
             open: htmlFile,
-        }
+        },
     };
 });
